@@ -1,6 +1,6 @@
 package de.htwdd.htwdresden.adapter.typeadapter;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonArray;
