@@ -1,6 +1,6 @@
 package de.htwdd.htwdresden.classes;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.Log;
 
 import java.util.Date;
