@@ -1,7 +1,7 @@
 package de.htwdd.htwdresden.classes;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.Log;
 
 import java.util.Calendar;
