@@ -2,8 +2,8 @@ package de.htwdd.htwdresden.classes;
 
 import android.content.Context;
 import android.content.SharedPreferences;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.widget.LinearLayout;
 
 import java.text.DateFormatSymbols;
