@@ -2,7 +2,7 @@ package de.htwdd.htwdresden.classes;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.widget.LinearLayout;
 
 import java.util.ArrayList;
