@@ -2,8 +2,8 @@ package de.htwdd.htwdresden.service;
 
 import android.app.IntentService;
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.util.Log;
 
 import de.htwdd.htwdresden.R;
