@@ -1,7 +1,7 @@
 package de.htwdd.htwdresden.types;
 
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import java.util.Date;
 
