@@ -1,6 +1,6 @@
 package de.htwdd.htwdresden.types.exams;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 /**
  * Statistik für Prüfungsleistungen
