@@ -1,7 +1,7 @@
 package de.htwdd.htwdresden.types.semsterPlan;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import java.util.Date;
 
